my-blog-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.115.12
fastapi-cache2==0.2.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
requests==2.32.3
cachetools==5.5.2
passlib==1.7.4
pendulum==3.0.0
pip==23.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
loguru==0.7.3
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
alembic==1.12.1
bcrypt==4.0.1
certifi==2023.7.22
dnspython==2.4.2
email-validator==2.1.0.post1
httpx==0.25.1
Mako==1.2.4
MarkupSafe==2.1.3
pytzdata==2020.1
charset-normalizer==3.4.1





